+++ packages/design-system/.storybook/ThemeDecorator.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
-import { ThemeProvider } from '@mui/material';
+import { ThemeProvider } from '@mui/material/styles';
 import { themes } from '../src/themes';
 import { useGlobals } from '@storybook/preview-api';